feat: extract products views to separated components
add description block to the product view with variations
