+++ client-app/pages/product/product.vue
@@ -9,111 +9,8 @@
         </button>
       </div>
       <h1 class="text-2xl md:text-4xl font-bold uppercase">{{ product.name }}</h1>
-      <div v-if="!withVariations && !isMobile" class="text-sm mt-1">
-        <span>Item #</span><span class="font-extrabold pl-1">{{ product.code }}</span>
-      </div>
-      <div class="flex flex-col md:flex-row md:space-x-6 mt-5">
-        <div
-          class="flex flex-grow flex-col lg:flex-row -mx-5 md:mx-0 lg:space-x-12 mb-6 p-6 bg-white border border-gray-100 rounded-md shadow-sm"
-        >
-          <div class="lg:w-1/3 mb-8 lg:mb-0">
-            <ImageGallery
-              :src="product.imgSrc ?? ''"
-              :images="product.images ?? []"
-              :is-mobile="isMobile"
-            ></ImageGallery>
-            <!-- Compare checkbox -->
-            <div class="mt-8 hidden md:flex items-center text-sm cursor-pointer">
-              <input
-                type="checkbox"
-                class="form-tick appearance-none w-5 h-5 border-2 border-gray-300 rounded-sm checked:bg-cyan-700 checked:border-transparent focus:outline-none cursor-pointer"
-              />
-              <span class="ml-2">Compare</span>
-            </div>
-          </div>
-          <div class="flex flex-col lg:w-2/3">
-            <ProductTitledBlock class="mt-5" image-src="/static/images/technical_specs.svg" title="technical specs">
-              <ProductProperties v-if="product.properties" :properties="product.properties"></ProductProperties>
-            </ProductTitledBlock>
-            <ProductTitledBlock
-              v-if="!withVariations && product?.description"
-              class="mt-5"
-              image-src="/static/images/description.svg"
-              title="Description"
-            >
-              <MarkdownRender :src="product?.description?.content" class="text-gray-500"></MarkdownRender>
-            </ProductTitledBlock>
-            <!-- variations  -->
-            <ProductTitledBlock
-              v-if="withVariations"
-              class="mt-5"
-              image-src="/static/images/variations_customize.svg"
-              title="Customize your order"
-            >
-              <ProductVariationCard class="mb-5" :variation="product"></ProductVariationCard>
-              <div v-for="(variation, i) in product.variations" :key="variation?.id ?? i">
-                <ProductVariationCard
-                  v-if="variation"
-                  class="mb-5"
-                  :variation="variation"
-                  :product="product"
-                ></ProductVariationCard>
-              </div>
-            </ProductTitledBlock>
-          </div>
-        </div>
-        <div class="flex-none md:w-80 lg:w-96 flex flex-col">
-          <div class="bg-white border shadow-sm rounded-md">
-            <div class="border-b px-5 py-4 md:px-7">
-              <h2 class="text-xl font-extrabold uppercase">Price &amp; Delivery</h2>
-            </div>
-            <div class="border-b px-5 py-8 md:px-7 md:py-5">
-              <!-- Product price -->
-              <div v-if="!withVariations" class="flex items-baseline justify-between text-sm">
-                <div class="font-extrabold text-base">Your price:</div>
-                <div>
-                  <span class="font-extrabold text-green-700">{{ product.price?.actual?.formattedAmount }}</span
-                  >&nbsp;<span class="font-semibold hidden lg:inline-block">/ each</span>
-                </div>
-              </div>
-              <div v-else class="flex items-baseline justify-between text-sm">
-                <div class="font-extrabold text-base">Total in your cart:</div>
-                <div class="font-extrabold">
-                  <span class="text-green-700">{{ currency?.symbol }}{{ variationsCartTotal.toFixed(2) }}</span>
-                </div>
-              </div>
-              <div class="mt-7 md:mt-5">
-                <AddToCart v-if="!withVariations" class="-mb-3" :product="product"></AddToCart>
-                <div v-else>
-                  <router-link
-                    class="block text-center bg-yellow-500 rounded text-white px-2 py-2 font-bold uppercase hover:bg-yellow-600 cursor-pointer"
-                    to="/checkout"
-                    >view cart</router-link
-                  >
-                </div>
-              </div>
-            </div>
-            <div class="flex text-center">
-              <div
-                class="flex items-center justify-center flex-1 py-4 px-1 border-r space-x-2 cursor-pointer hover:bg-gray-100"
-              >
-                <i class="fas fa-plus text-base text-yellow-500"></i>
-                <span class="text-sm text-blue-800 font-bold">Add to list</span>
-              </div>
-              <div
-                class="flex items-center justify-center flex-1 py-4 px-1 border-r space-x-2 cursor-pointer hover:bg-gray-100"
-              >
-                <i class="fas fa-envelope fa-xl text-yellow-500"></i
-                ><span class="text-sm text-blue-800 font-bold">Email</span>
-              </div>
-              <div class="flex items-center justify-center flex-1 py-4 px-1 space-x-2 cursor-pointer hover:bg-gray-100">
-                <i class="fas fa-print text-yellow-500"></i>
-                <span class="text-sm text-blue-800 font-bold" @click="print()">Print</span>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
+      <ProductCard v-if="!withVariations" :product="product"></ProductCard>
+      <ProductWithVariationsCard v-else :product="product"></ProductWithVariationsCard>
     </div>
   </div>
 </template>
@@ -126,23 +23,15 @@ import {
   useBreadcrumbs,
   Breadcrumbs,
   IBreadcrumbsItem,
-  ProductProperties,
-  ProductTitledBlock,
-  ProductVariationCard,
+  ProductCard,
+  ProductWithVariationsCard,
 } from "@/shared/catalog";
-import { AddToCart, useCart } from "@/shared/cart";
-import { MarkdownRender, ImageGallery } from "@/components";
-import { useBreakpoints, breakpointsTailwind } from "@vueuse/core";
 import { Breadcrumb } from "@/core/api/graphql/types";
-
-const breakpoints = useBreakpoints(breakpointsTailwind);
-const isMobile = breakpoints.smaller("lg");
 
 const route = useRoute();
 const router = useRouter();
 
-const { product, loading, loadProduct, withVariations, variationsCartTotal } = useProducts();
-const { currency } = useCart();
+const { product, loading, loadProduct, withVariations } = useProducts();
 
 const breadcrumbsItems: Ref<IBreadcrumbsItem[]> = ref([{ url: "/", title: "Home" }]);
 const { buildBreadcrumbs } = useBreadcrumbs();
@@ -153,10 +42,6 @@ onMounted(async () => {
   await loadProduct(productId.value);
   breadcrumbsItems.value = buildBreadcrumbs(product.value.breadcrumbs as Breadcrumb[]);
 });
-
-function print() {
-  window.print();
-}
 
 function back() {
   router.back();
